+++ pkg/resource/events-logger-config/reconciler.go
@@ -20,6 +20,8 @@ import (
 // EventsLogger config: extra events-logger config defining what we want to retrieve.
 type Reconciler struct {
 	client.Client
+	IncludeNamespaces []string
+	ExcludeNamespaces []string
 }
 
 // ReconcileCreate ensures events-logger config is created with the right credentials
@@ -28,7 +30,7 @@ func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Inter
 	logger.Info("events-logger-config create")
 
 	// Get desired config
-	desiredEventsLoggerConfig, err := generateEventsLoggerConfig(lc)
+	desiredEventsLoggerConfig, err := generateEventsLoggerConfig(lc, r.IncludeNamespaces, r.ExcludeNamespaces)
 	if err != nil {
 		logger.Info("events-logger-config - failed generating events-logger config!", "error", err)
 		return ctrl.Result{}, errors.WithStack(err)