+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Collect all cluster events from workload clusters.
+- Adds support to include and exclude event collection per namespace in workload clusters. If nothing is configured, the event collector will collect all events in the WC.
 
 ## [0.24.1] - 2025-03-06